engine build ok

+++ core/src/wrapper/BinVecImpl.h
@@ -46,9 +46,14 @@ class BinVecImpl : public VecIndexImpl {
     Status
     SetBlacklist(faiss::ConcurrentBitsetPtr list) override;
 
+    //    Status
+    //    SearchById(const int64_t& nq, const uint8_t* xq, faiss::ConcurrentBitsetPtr bitset, float* dist, int64_t* ids,
+    //               const Config& cfg) override;
+
     Status
-    SearchById(const int64_t& nq, const uint8_t* xq, faiss::ConcurrentBitsetPtr bitset, float* dist, int64_t* ids,
+    SearchById(const int64_t& nq, const int64_t* xq, faiss::ConcurrentBitsetPtr bitset, float* dist, int64_t* ids,
                const Config& cfg) override;
+
     VecIndexPtr
     CopyToCpu(const Config& cfg) override;
 };
